omit empty input lines from output, and swap header barcodes around a bit
Because padding has to be done for the error correction, a potentially
large number of empty lines could result on a sheet. This would waste
lots of ink and provide no advantage at all.

So these lines are just not recorded, and the fact that they are not
included is recorded in another barcode at the top of the file.