decoder: properly truncate output data
Sometimes a page isn't completely filled with data. In that case,
it is filled with null bytes (or has recorded in the header that it
would be). To prevent these null bytes from being appended to the
data, decoder now uses the information from the header to truncate
output properly.